Add files via upload
I have an initial portfolio with short positions Asian Call options on the SXO and short digital put options on Barricks Gold. I was required to hedge my portfolio with a mix of vanilla options and the underliers. "Portfolio Trading and Hedging" will have all the positions throughout the 9 weeks of this project. At the end I calculated risk measures such as standard deviation, VaR and CVaR on simulations of the worst week of the portfolio. These simulations were generated by the CreditMetrics Model and Bootstrap Methods on the past 12 months of data.